backend/requirements.txt
asgiref==3.6.0
Django==4.2
sqlparse==0.4.3
tzdata==2023.3
django-phonenumber-field==6.1.0
phonenumbers==8.12.50
Pillow==9.1.0
